unify the file format under cmd, alluxio jindo dataset
