+++ cmd/dataset/app/init.go
@@ -14,19 +14,18 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package app
 
-import (
-	"fmt"
-	"os"
+import "github.com/spf13/cobra"
 
-	"github.com/fluid-cloudnative/fluid/cmd/dataset/app"
-)
-
-func main() {
-	command := app.NewDatasetCommand()
-	if err := command.Execute(); err != nil {
-		fmt.Fprintf(os.Stderr, "%s", err.Error())
-		os.Exit(1)
+func NewDatasetCommand() *cobra.Command {
+	command := &cobra.Command{
+		Use:   "dataset-controller",
+		Short: "controller for dataset",
 	}
+
+	command.AddCommand(versionCmd, datasetCmd)
+
+	return command
+
 }
