+++ cmd/jindo/app/jindo.go
@@ -1,0 +1,134 @@
+/*
+Copyright 2021 The Fluid Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package app
+
+import (
+	"os"
+
+	"github.com/fluid-cloudnative/fluid"
+	datav1alpha1 "github.com/fluid-cloudnative/fluid/api/v1alpha1"
+	jindoctl "github.com/fluid-cloudnative/fluid/pkg/controllers/v1alpha1/jindo"
+	"github.com/fluid-cloudnative/fluid/pkg/ddc/base"
+	"github.com/fluid-cloudnative/fluid/pkg/ddc/base/portallocator"
+	"github.com/fluid-cloudnative/fluid/pkg/ddc/jindo"
+	"github.com/fluid-cloudnative/fluid/pkg/utils"
+	"github.com/spf13/cobra"
+	zapOpt "go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/util/net"
+	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/controller"
+	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+)
+
+var (
+	scheme   = runtime.NewScheme()
+	setupLog = ctrl.Log.WithName("setup")
+	// Use compiler to check if the struct implements all the interface
+	_ base.Implement = (*jindo.JindoEngine)(nil)
+
+	metricsAddr          string
+	enableLeaderElection bool
+	development          bool
+	// The new mode
+	eventDriven             bool
+	portRange               string
+	maxConcurrentReconciles int
+	pprofAddr               string
+)
+
+var jindoCmd = &cobra.Command{
+	Use:   "start",
+	Short: "start jindoruntime-controller in Kubernetes",
+	Run: func(cmd *cobra.Command, args []string) {
+		handle()
+	},
+}
+
+func init() {
+	_ = clientgoscheme.AddToScheme(scheme)
+	_ = datav1alpha1.AddToScheme(scheme)
+
+	jindoCmd.Flags().StringVarP(&metricsAddr, "metrics-addr", "", ":8080", "The address the metric endpoint binds to.")
+	jindoCmd.Flags().BoolVarP(&enableLeaderElection, "enable-leader-election", "", false, "Enable leader election for controller manager. Enabling this will ensure there is only one active controller manager.")
+	jindoCmd.Flags().BoolVarP(&development, "development", "", true, "Enable development mode for fluid controller.")
+	jindoCmd.Flags().StringVar(&portRange, "runtime-node-port-range", "18000-19999", "Set available port range for Jindo")
+	jindoCmd.Flags().IntVar(&maxConcurrentReconciles, "runtime-workers", 3, "Set max concurrent workers for JindoRuntime controller")
+	jindoCmd.Flags().BoolVar(&eventDriven, "event-driven", true, "The reconciler's loop strategy. if it's false, it indicates period driven.")
+	jindoCmd.Flags().StringVarP(&pprofAddr, "pprof-addr", "", "", "The address for pprof to use while exporting profiling results")
+}
+
+func handle() {
+	fluid.LogVersion()
+
+	ctrl.SetLogger(zap.New(func(o *zap.Options) {
+		o.Development = development
+	}, func(o *zap.Options) {
+		o.ZapOpts = append(o.ZapOpts, zapOpt.AddCaller())
+	}, func(o *zap.Options) {
+		if !development {
+			encCfg := zapOpt.NewProductionEncoderConfig()
+			encCfg.EncodeLevel = zapcore.CapitalLevelEncoder
+			encCfg.EncodeTime = zapcore.ISO8601TimeEncoder
+			o.Encoder = zapcore.NewConsoleEncoder(encCfg)
+		}
+	}))
+
+	utils.NewPprofServer(setupLog, pprofAddr)
+
+	mgr, err := ctrl.NewManager(ctrl.GetConfigOrDie(), ctrl.Options{
+		Scheme:             scheme,
+		MetricsBindAddress: metricsAddr,
+		LeaderElection:     enableLeaderElection,
+		LeaderElectionID:   "5688274864.data.fluid.io",
+		Port:               9443,
+	})
+	if err != nil {
+		setupLog.Error(err, "unable to start jindoruntime manager")
+		os.Exit(1)
+	}
+
+	controllerOptions := controller.Options{
+		MaxConcurrentReconciles: maxConcurrentReconciles,
+	}
+
+	if err = (jindoctl.NewRuntimeReconciler(mgr.GetClient(),
+		ctrl.Log.WithName("jindoctl").WithName("JindoRuntime"),
+		mgr.GetScheme(),
+		mgr.GetEventRecorderFor("JindoRuntime"),
+	)).SetupWithManager(mgr, controllerOptions, eventDriven); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "JindoRuntime")
+		os.Exit(1)
+	}
+
+	pr, err := net.ParsePortRange(portRange)
+	if err != nil {
+		setupLog.Error(err, "can't parse port range. Port range must be like <min>-max")
+		os.Exit(1)
+	}
+	setupLog.Info("port range parsed", "port range", pr.String())
+
+	portallocator.SetupRuntimePortAllocator(mgr.GetClient(), pr, jindo.GetReservedPorts)
+
+	setupLog.Info("starting jindoruntime-controller")
+	if err := mgr.Start(ctrl.SetupSignalHandler()); err != nil {
+		setupLog.Error(err, "problem jindoruntime-controller")
+		os.Exit(1)
+	}
+}
